+++ src/reconstruction/index.ts
@@ -12,3 +12,4 @@ export * from "./identity";
 export * from "./packageV2";
 export * from "./spatialV2";
 export * from "./types";
+export * from "./caoDomain";